FE: Change export strategy of Google Login script
+ Turn Google login callback as component property
